requirements.txt
fastapi
uvicorn
python-redmine
openpyxl
pydantic
schedule
python-multipart
email-validator
python-dotenv
apscheduler
pandas